set aside original local scripts, building alts as part of multipart app
